added event search
Added a separate html page for event searching. This is currently experimental and will be added into the main page later on. It does not pool actual event titles and dates since the event object structure hasn't been defined yet, but test data is pooled in to show that the basic functionality works.