Ported functions/Added modal display
I ported functions for the editing/deleting part of the events to a new file since it is for the most part the same as searchevents. I also modified how to get the eventID for editing since we aren't visually displaying it anymore.

The event data gets prepopulated in the modal on the main page, but the modal is still a bit wonky due to trying to mesh together two types of modals. Will work on fixing that.